# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VA eBay Client Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wise-cards:
    get:
      operationId: getWiseCards
      tags: [wise-cards]
      summary: List all Wise cards
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WiseCard"
    post:
      operationId: createWiseCard
      tags: [wise-cards]
      summary: Create Wise card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWiseCardBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WiseCard"
  /wise-cards/{id}:
    patch:
      operationId: updateWiseCard
      tags: [wise-cards]
      summary: Update Wise card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWiseCardBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WiseCard"
    delete:
      operationId: deleteWiseCard
      tags: [wise-cards]
      summary: Delete Wise card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Bank Accounts ──────────────────────────────────────────────────────────
...
  /violations:
    get:
      operationId: getViolations
      tags: [violations]
      summary: List all violations
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"
    post:
      operationId: createViolation
      tags: [violations]
      summary: Create violation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateViolationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
  /violations/{id}:
    patch:
      operationId: updateViolation
      tags: [violations]
      summary: Update violation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateViolationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
    delete:
      operationId: deleteViolation
      tags: [violations]
      summary: Delete violation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tasks ──────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Earnings ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WiseCard:
      type: object
      properties:
        id:
          type: integer
        cardCode:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        provider:
          type: ["string", "null"]
        cardType:
          type: ["string", "null"]
        cardHolder:
          type: ["string", "null"]
        cardNumberMasked:
          type: ["string", "null"]
        expiryMonth:
          type: ["integer", "null"]
        expiryYear:
          type: ["integer", "null"]
        currency:
          type: ["string", "null"]
        wiseEmail:
          type: ["string", "null"]
        wisePassword:
          type: ["string", "null"]
        wise2fa:
          type: boolean
        linkedBankId:
          type: ["string", "null"]
        balance:
          type: ["number", "null"]
        daysToExpiry:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, cardCode, clientId, clientName, wise2fa, createdAt, updatedAt]
    CreateWiseCardBody:
      type: object
      properties:
        cardCode:
          type: string
        clientId:
          type: integer
        provider:
          type: ["string", "null"]
        cardType:
          type: ["string", "null"]
        cardHolder:
          type: ["string", "null"]
        cardNumberMasked:
          type: ["string", "null"]
        expiryMonth:
          type: ["integer", "null"]
        expiryYear:
          type: ["integer", "null"]
        currency:
          type: ["string", "null"]
        wiseEmail:
          type: ["string", "null"]
        wisePassword:
          type: ["string", "null"]
        wise2fa:
          type: boolean
        linkedBankId:
          type: ["string", "null"]
        balance:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [cardCode, clientId, wise2fa]
    UpdateWiseCardBody:
      type: object
      properties:
        cardCode:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        provider:
          type: ["string", "null"]
        cardType:
          type: ["string", "null"]
        cardHolder:
          type: ["string", "null"]
        cardNumberMasked:
          type: ["string", "null"]
        expiryMonth:
          type: ["integer", "null"]
        expiryYear:
          type: ["integer", "null"]
        currency:
          type: ["string", "null"]
        wiseEmail:
          type: ["string", "null"]
        wisePassword:
          type: ["string", "null"]
        wise2fa:
          type: ["boolean", "null"]
        linkedBankId:
          type: ["string", "null"]
        balance:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Violation:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        ebayUsername:
          type: string
        date:
          type: ["string", "null"]
        policyCode:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        actionTaken:
          type: ["string", "null"]
        resolved:
          type: boolean
        resolvedOn:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, clientName, ebayUsername, resolved, createdAt, updatedAt]
    CreateViolationBody:
      type: object
      properties:
        clientId:
          type: integer
        ebayUsername:
          type: string
        date:
          type: ["string", "null"]
        policyCode:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        actionTaken:
          type: ["string", "null"]
        resolved:
          type: boolean
        resolvedOn:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [clientId, ebayUsername, resolved]
    UpdateViolationBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        ebayUsername:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        policyCode:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        actionTaken:
          type: ["string", "null"]
        resolved:
          type: ["boolean", "null"]
        resolvedOn:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Task:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        task:
          type: string
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        daysLeft:
          type: ["integer", "null"]
        done:
          type: boolean
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, task, done, status, createdAt, updatedAt]
    CreateTaskBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        task:
          type: string
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        done:
          type: boolean
        status:
          type: string
        notes:
          type: ["string", "null"]
      required: [task, done, status]
    UpdateTaskBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        task:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        done:
          type: ["boolean", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...